add parser.section.case_information
